backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
werkzeug==2.2.3
databricks-sdk==0.23.0
python-dotenv==0.21.0
pydantic==1.10.7
